CachedList 0.1
Accessing (getting, setting) list values like a regular array works as
expected. If the value is not in the currently loaded part of the list,
then the part required is swapped in its place.

The main issue so far is that it does not support more than one cached
list at a time as they will clash and overwrite each other's data. This
will be fixed in future releases.